Add GNU Affero General Public License version 3 to the project.
This commit adds the GNU Affero General Public License version 3 to the
codebase. The GNU Affero General Public License is designed to ensure
cooperation with the community in the case of network server software
and guarantees users the freedom to share and change all versions of the
program. The license requires the operator of a network server to
provide the source code of any modified version to the users of that
server.